Scale sampled data properly
Needs to be rescaled back to int for storage, 8* higher resolution so the hw can use oversampling.

BLDC can use floats for these variables, since they get changed to floats later anyway. Might as well do it earlier to make the code simpler.

+++ motor/mcpwm.c
@@ -1457,14 +1457,14 @@ static THD_FUNCTION(timer_thread, arg) {
 void mcpwm_adc_inj_int_handler(void) {
 	uint32_t t_start = timer_time_now();
 
-	int curr0 = HW_GET_INJ_CURR1();
-	int curr1 = HW_GET_INJ_CURR2();
+	float curr0 = HW_GET_INJ_CURR1();
+	float curr1 = HW_GET_INJ_CURR2();
 
-	int curr0_2 = HW_GET_INJ_CURR1_S2();
-	int curr1_2 = HW_GET_INJ_CURR1_S2();
+	float curr0_2 = HW_GET_INJ_CURR1_S2();
+	float curr1_2 = HW_GET_INJ_CURR1_S2();
 
 #ifdef HW_HAS_3_SHUNTS
-	int curr2 = HW_GET_INJ_CURR3();
+	float curr2 = HW_GET_INJ_CURR3();
 #endif
 
 #ifdef INVERTED_SHUNT_POLARITY
